Fixed memory glitch
Removed the part that checks whether the node already has k diverse
paths before adding paths to it. For some reason, that was making it eat
up all the memory. (Future research - figure out why that happens, see
if there's a way to make that optimization without glitching the
system.)